adds the antd table to the goals
