+++ client/src/App.css
@@ -1,4 +1,4 @@
-@import '~antd/dist/antd.css';
+@import '~antd/dist/antd.less';
 
 * {
   margin: 0;
