--- client/src/features/commonComponents/GoalList.js
@@ -1,22 +1,0 @@
-import React, { useEffect } from 'react'
-import { useDispatch } from 'react-redux'
-import { List, Avatar, Space } from 'antd'
-import GoalItem from './goals/GoalItem'
-import { deleteGoalById } from '../common/goalSlice'
-// import { getGoalsByChildId } from './goals/goalSlice'
-
-export default function GoalList(props) {
-  const dispatch = useDispatch()
-  return (
-    <>
-      {props.goals.map((goal) => (
-        <GoalItem
-          key={'goal-' + goal.id}
-          goal={goal}
-          onDelete={(id) => dispatch(deleteGoalById(id))}
-          onStatusChange={props.onStatusChange}
-        />
-      ))}
-    </>
-  )
-}
