+++ client/src/features/parent/ChildTab.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import { Tabs } from 'antd'
 import AddGoalModal from '../parent/AddGoalModal'
 import AddPrizeModal from '../parent/AddPrizeModal'
-import GoalList from '../commonComponents/GoalList'
+import GoalList from '../commonComponents/goals/GoalList'
 import PrizesList from '../commonComponents/PrizesList'
 import styles from '../parent/Tabs.module.css'
 import AddChild from './AddChildModal'
