+++ client/src/features/sandbox/Sandbox.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { Tabs } from 'antd'
 import AddGoalModal from '../parent/AddGoalModal'
 import AddPrizeModal from '../parent/AddPrizeModal'
-import GoalList from '../commonComponents/GoalList'
+import GoalList from '../commonComponents/goals/GoalList'
 import PrizesList from '../commonComponents/PrizesList'
 import styles from '../parent/Tabs.module.css'
 // import ModalTest from '../parent/AddChildModal'
